+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added
   - Support for workflow IDs and attributes (used by Forge Webhooks) in the `Translate Object (Custom)` command.
+- Fixed
+  - Issue with creating a webhook with no attribute data.
 
 ## [2.2.0] - 2020-10-23